[fix] Do not attempt to create symlink if already present

+++ install-arduino-sktech.sh
@@ -4,12 +4,14 @@
 # to this repository.
 
 mkdir -p $HOME/Arduino/ZumoPdS
-src=
-ln -s $PWD/arduino/src/sketch.ino $HOME/Arduino/ZumoPdS/ZumoPdS.ino
+src=$PWD/arduino/src/sketch.ino
+dst=$HOME/Arduino/ZumoPdS/ZumoPdS.ino
+
+[[ -e $dst ]] || ln -s $src $dst
 ls -l --color $HOME/Arduino/ZumoPdS/ZumoPdS.ino
 
-for src in $PWD/arduino/lib/Zumo*; do
+for src in $PWD/arduino/lib/Zumo* $PWD/arduino/lib/QTRSensors; do
     dst=$HOME/Arduino/libraries/$(basename $src)
-    ln -s $src $dst 
+    [[ -e $dst ]] || ln -s $src $dst 
     ls -l --color $dst
 done
